pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-cfi"
version = "0.1.0"
description = "Control Flow Integrity for AI agents. CI gate that detects prompt-injection-induced tool-graph drift."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "agent-cfi contributors" }]
keywords = ["ai-security", "prompt-injection", "llm", "agents", "sdlc", "ci", "cfi", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "networkx>=3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]

[project.scripts]
agent-cfi = "agent_cfi.cli:main"

[project.urls]
Homepage = "https://github.com/grcwarlock/agent-cfi"
Issues = "https://github.com/grcwarlock/agent-cfi/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
agent_cfi = ["probes/*.yaml"]
